# openapi.yaml
openapi: "3.0.3"
info:
  title: Printing Press Manager API
  version: "1.0.0"
  description: |
    Backend server for a kiosk-based printing press system.
    Manages kiosks, documents, jobs, payments, and print commands.
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
            requestId:
              type: string
              nullable: true
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: string
        tokenType:
          type: string
...
    Document:
      type: object
      properties:
        id:
          type: string
        kioskId:
          type: string
        userPhone:
          type: string
          nullable: true
        fileUrl:
          type: string
        fileName:
          type: string
        pageCount:
          type: integer
        checksum:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      summary: Login with username/password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Refreshed tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      summary: Logout and revoke refresh token
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /documents:
    post:
      summary: Upload document metadata (chatbot sends this)
      tags: [Documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kioskId, fileUrl, pageCount]
              properties:
                kioskId:
                  type: string
                userPhone:
                  type: string
                fileUrl:
                  type: string
                  description: URL where document can be downloaded
                fileName:
                  type: string
                pageCount:
                  type: integer
                checksum:
                  type: string
                  description: SHA-256 or MD5 checksum of file
      responses:
        "201":
          description: Document stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{docId}:
    get:
      summary: Get document metadata
      tags: [Documents]
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...